Set screen capture for Diablo
Diablo in a windowed mode always starts up in the middlfe of the screen.
This is not really nice, but ofr now its okay.
Assuming Diablo runs in 640x480, on a fullHD monitor,
the current capture should be fine.
Later on this can be improved. Maybe skipping the imShow totally,
and using fullscreen.