Add Dockerfile for Tesseract testing
I want to automate the testing of Tesseract.
Installing on Linux can be automated
a lot easier than on Windows,
so I need a Linux host for that.
Also the build host has to be Windows for the win32 tools.

That leaves Docker as a viable way to go.
This way the testing environment will not be 100% prod like,
but this difference is probably okay.